requirements.txt
flask
ultralytics
opencv-python-headless
numpy
pillow
easyocr
requests
torch
